+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Add experimental `Restify::Adapter::Ethon` adapter directly using libcurl in a background event loop
 - Perf: Run event loop in waiting threads to avoid slow handover between threads
+- Add `restify/webmock` to intercept requests with WebMock for all adapters
 
 ### Changes